@@ -3,22 +3,21 @@ package es.upsa.mimo.gamercollection.adapters
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
-import es.upsa.mimo.gamercollection.databinding.SongItemBinding
+import es.upsa.mimo.gamercollection.databinding.ItemSongBinding
 import es.upsa.mimo.gamercollection.models.responses.SongResponse
-import es.upsa.mimo.gamercollection.utils.Constants
+import es.upsa.mimo.gamercollection.viewholders.SongsViewHolder
 
 class SongsAdapter(
     private var songs: List<SongResponse>,
     private var editable: Boolean,
     private var onItemClickListener: OnItemClickListener
-) : RecyclerView.Adapter<SongsAdapter.SongsViewHolder?>() {
+) : RecyclerView.Adapter<SongsViewHolder?>() {
 
     //region Lifecycle methods
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SongsViewHolder {
         return SongsViewHolder(
-            SongItemBinding.inflate(
+            ItemSongBinding.inflate(
                 LayoutInflater.from(parent.context),
                 parent,
                 false
@@ -31,7 +30,7 @@ class SongsAdapter(
     }
 
     override fun onBindViewHolder(holder: SongsViewHolder, position: Int) {
-        holder.bind(songs[position])
+        holder.bind(songs[position], editable, onItemClickListener)
     }
     //endregion
 
@@ -50,37 +49,4 @@ class SongsAdapter(
         notifyDataSetChanged()
     }
     //endregion
-
-    inner class SongsViewHolder(val binding: SongItemBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-
-        fun bind(_song: SongResponse) {
-            with(binding) {
-
-                darkMode = Constants.isDarkMode(root.context)
-                song = _song
-                editable = this@SongsAdapter.editable
-                onItemClickListener = this@SongsAdapter.onItemClickListener
-                executePendingBindings()
-            }
-        }
-    }
-}
-
-@BindingAdapter("songs")
-fun setRecyclerViewSongs(recyclerView: RecyclerView?, songs: List<SongResponse>?) {
-
-    val adapter = recyclerView?.adapter
-    if (adapter is SongsAdapter && songs != null) {
-        adapter.setSongs(songs)
-    }
-}
-
-@BindingAdapter("editable")
-fun setRecyclerViewEditable(recyclerView: RecyclerView?, editable: Boolean?) {
-
-    val adapter = recyclerView?.adapter
-    if (adapter is SongsAdapter && editable != null) {
-        adapter.setEditable(editable)
-    }
 }
