@@ -1,6 +1,5 @@
 package es.upsa.mimo.gamercollection.base
 
-import android.app.AlertDialog
 import android.app.SearchManager
 import android.content.Context
 import android.content.Intent
@@ -11,18 +10,24 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.SearchView
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.AppCompatImageView
+import androidx.appcompat.widget.Toolbar
 import androidx.core.content.ContextCompat
 import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentTransaction
 import androidx.lifecycle.MutableLiveData
+import androidx.navigation.fragment.findNavController
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import es.upsa.mimo.gamercollection.R
-import es.upsa.mimo.gamercollection.fragments.popups.PopupErrorDialogFragment
+import es.upsa.mimo.gamercollection.extensions.isDarkMode
+import es.upsa.mimo.gamercollection.extensions.setStatusBarStyle
 import es.upsa.mimo.gamercollection.fragments.popups.PopupLoadingDialogFragment
 import es.upsa.mimo.gamercollection.fragments.popups.PopupSyncAppDialogFragment
 import es.upsa.mimo.gamercollection.models.responses.ErrorResponse
 import es.upsa.mimo.gamercollection.utils.Constants
+import es.upsa.mimo.gamercollection.utils.StatusBarStyle
 import java.io.Serializable
 import java.lang.reflect.ParameterizedType
 
@@ -39,6 +44,9 @@ abstract class BindingFragment<Binding : ViewDataBinding> : Fragment() {
     //region Protected properties
     protected lateinit var binding: Binding
         private set
+    protected abstract val statusBarStyle: StatusBarStyle
+    protected abstract val hasOptionsMenu: Boolean
+    protected open var toolbar: Toolbar? = null
     //endregion
 
     //region Lifecycle methods
@@ -47,6 +55,8 @@ abstract class BindingFragment<Binding : ViewDataBinding> : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
+        setHasOptionsMenu(hasOptionsMenu)
         val bindingType =
             (this.javaClass.genericSuperclass as ParameterizedType).actualTypeArguments
                 .firstOrNull {
@@ -67,20 +77,55 @@ abstract class BindingFragment<Binding : ViewDataBinding> : Fragment() {
         binding = inflateMethod.invoke(null, inflater, container, false) as Binding
         return binding.root
     }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        activity?.let {
+            when (statusBarStyle) {
+                StatusBarStyle.PRIMARY -> {
+                    it.window.setStatusBarStyle(
+                        ContextCompat.getColor(it, R.color.colorSecondary),
+                        !it.isDarkMode()
+                    )
+                }
+                StatusBarStyle.SECONDARY -> {
+                    it.window.setStatusBarStyle(
+                        ContextCompat.getColor(it, R.color.colorPrimary),
+                        it.isDarkMode()
+                    )
+                }
+            }
+        }
+    }
+    //endregion
+
+    //region Protected methods
+    protected open fun initializeUi() {
+
+        toolbar?.let {
+            (activity as? AppCompatActivity)?.setSupportActionBar(it)
+            it.setNavigationOnClickListener {
+                findNavController().popBackStack()
+            }
+        }
+    }
     //endregion
 
     //region Public methods
     fun showPopupDialog(message: String, goBack: MutableLiveData<Boolean>? = null) {
 
-        val ft: FragmentTransaction = activity?.supportFragmentManager?.beginTransaction() ?: return
-        val prev = activity?.supportFragmentManager?.findFragmentByTag(Constants.POPUP_DIALOG)
-        if (prev != null) {
-            ft.remove(prev)
-        }
-        ft.addToBackStack(null)
-        val dialogFragment = PopupErrorDialogFragment(message, goBack)
-        dialogFragment.isCancelable = false
-        dialogFragment.show(ft, Constants.POPUP_DIALOG)
+        MaterialAlertDialogBuilder(requireContext())
+            .setMessage(message)
+            .setCancelable(false)
+            .setPositiveButton(resources.getString(R.string.accept)) { dialog, _ ->
+
+                dialog.dismiss()
+                goBack?.let {
+                    it.value = true
+                }
+            }
+            .show()
     }
 
     fun showLoading() {
@@ -116,9 +161,13 @@ abstract class BindingFragment<Binding : ViewDataBinding> : Fragment() {
         showPopupDialog(error.toString())
     }
 
-    fun showPopupConfirmationDialog(message: String, acceptHandler: () -> Unit, cancelHandler: (() -> Unit)? = null) {
+    fun showPopupConfirmationDialog(
+        message: String,
+        acceptHandler: () -> Unit,
+        cancelHandler: (() -> Unit)? = null
+    ) {
 
-        AlertDialog.Builder(context)
+        MaterialAlertDialogBuilder(requireContext())
             .setMessage(message)
             .setCancelable(false)
             .setPositiveButton(resources.getString(R.string.accept)) { dialog, _ ->
@@ -132,17 +181,27 @@ abstract class BindingFragment<Binding : ViewDataBinding> : Fragment() {
             .show()
     }
 
-    fun <T> launchActivity(activity: Class<T>) {
+    fun <T> launchActivity(activity: Class<T>, clearStack: Boolean = false) {
 
-        val intent = Intent(context, activity).apply {}
+        val intent = Intent(context, activity)
+        if (clearStack) {
+            intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+        }
         startActivity(intent)
     }
 
-    fun <T> launchActivityWithExtras(activity: Class<T>, params: Map<String, Serializable>) {
+    fun <T> launchActivityWithExtras(
+        activity: Class<T>,
+        params: Map<String, Serializable>,
+        clearStack: Boolean = false
+    ) {
 
-        val intent = Intent(context, activity).apply {}
+        val intent = Intent(context, activity)
         for (param in params) {
             intent.putExtra(param.key, param.value)
+        }
+        if (clearStack) {
+            intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
         }
         startActivity(intent)
     }
