@@ -1,25 +1,33 @@
 package es.upsa.mimo.gamercollection.fragments
 
 import android.os.Bundle
-import android.view.*
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.View
 import android.widget.SearchView
-import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.ObservableField
 import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import es.upsa.mimo.gamercollection.R
-import es.upsa.mimo.gamercollection.activities.GameDetailActivity
 import es.upsa.mimo.gamercollection.adapters.GamesAdapter
 import es.upsa.mimo.gamercollection.adapters.OnItemClickListener
 import es.upsa.mimo.gamercollection.base.BindingFragment
 import es.upsa.mimo.gamercollection.databinding.FragmentGameSearchBinding
-import es.upsa.mimo.gamercollection.fragments.GamesFragment.ScrollPosition
+import es.upsa.mimo.gamercollection.extensions.hideSoftKeyboard
 import es.upsa.mimo.gamercollection.utils.Constants
+import es.upsa.mimo.gamercollection.utils.ScrollPosition
+import es.upsa.mimo.gamercollection.utils.StatusBarStyle
 import es.upsa.mimo.gamercollection.viewmodelfactories.GameSearchViewModelFactory
 import es.upsa.mimo.gamercollection.viewmodels.GameSearchViewModel
 
 class GameSearchFragment : BindingFragment<FragmentGameSearchBinding>(), OnItemClickListener {
+
+    //region Protected properties
+    override val statusBarStyle = StatusBarStyle.SECONDARY
+    override val hasOptionsMenu = true
+    //endregion
 
     //region Private properties
     private lateinit var viewModel: GameSearchViewModel
@@ -28,17 +36,11 @@ class GameSearchFragment : BindingFragment<FragmentGameSearchBinding>(), OnItemC
     //endregion
 
     //region Lifecycle methods
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        setHasOptionsMenu(true)
-        return super.onCreateView(inflater, container, savedInstanceState)
-    }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        initializeUI()
+
+        toolbar = binding.toolbar
+        initializeUi()
     }
 
     override fun onResume() {
@@ -63,8 +65,8 @@ class GameSearchFragment : BindingFragment<FragmentGameSearchBinding>(), OnItemC
     //region Interface methods
     override fun onItemClick(id: Int) {
 
-        val params = mapOf(Constants.GAME_ID to id, Constants.IS_RAWG_GAME to true)
-        launchActivityWithExtras(GameDetailActivity::class.java, params)
+        val action = GameSearchFragmentDirections.actionSearchFragmentToGameDetailFragment(id, true)
+        findNavController().navigate(action)
     }
 
     override fun onSubItemClick(id: Int) {
@@ -98,14 +100,15 @@ class GameSearchFragment : BindingFragment<FragmentGameSearchBinding>(), OnItemC
     }
     //endregion
 
-    //region Private methods
-    private fun initializeUI() {
+    //region Protected methods
+    override fun initializeUi() {
+        super.initializeUi()
 
         val application = activity?.application
         viewModel = ViewModelProvider(
             this,
             GameSearchViewModelFactory(application)
-        ).get(GameSearchViewModel::class.java)
+        )[GameSearchViewModel::class.java]
         setupBindings()
 
         with(binding) {
@@ -123,7 +126,6 @@ class GameSearchFragment : BindingFragment<FragmentGameSearchBinding>(), OnItemC
 
             gamesAdapter = GamesAdapter(
                 this@GameSearchFragment.viewModel.games.value ?: listOf(),
-                this@GameSearchFragment.viewModel.platforms,
                 null,
                 this@GameSearchFragment
             )
@@ -156,39 +158,27 @@ class GameSearchFragment : BindingFragment<FragmentGameSearchBinding>(), OnItemC
 
         scrollPosition.set(ScrollPosition.TOP)
     }
+    //endregion
 
+    //region Protected methods
     private fun setupBindings() {
 
-        viewModel.gamesLoading.observe(viewLifecycleOwner, { isLoading ->
+        viewModel.gamesLoading.observe(viewLifecycleOwner) { isLoading ->
 
             if (isLoading) {
                 showLoading()
             } else {
                 hideLoading()
             }
-        })
+        }
 
-        viewModel.gamesError.observe(viewLifecycleOwner, { error ->
+        viewModel.gamesError.observe(viewLifecycleOwner) { error ->
             manageError(error)
-        })
+        }
 
-        viewModel.gamesCount.observe(viewLifecycleOwner, {
-            setTitle(it)
-        })
-
-        viewModel.scrollPosition.observe(viewLifecycleOwner, {
+        viewModel.scrollPosition.observe(viewLifecycleOwner) {
             scrollPosition.set(it)
-        })
-    }
-
-    private fun setTitle(gamesCount: Int) {
-
-        val title = resources.getQuantityString(
-            R.plurals.games_number_title,
-            gamesCount,
-            Constants.getFormattedNumber(gamesCount)
-        )
-        (activity as AppCompatActivity?)?.supportActionBar?.title = title
+        }
     }
 
     private fun reset() {
@@ -226,7 +216,7 @@ class GameSearchFragment : BindingFragment<FragmentGameSearchBinding>(), OnItemC
 
         viewModel.query = query
         reset()
-        Constants.hideSoftKeyboard(requireActivity())
+        requireActivity().hideSoftKeyboard()
     }
     //endregion
 }
