@@ -7,14 +7,20 @@ import es.upsa.mimo.gamercollection.R
 import es.upsa.mimo.gamercollection.activities.MainActivity
 import es.upsa.mimo.gamercollection.base.BindingFragment
 import es.upsa.mimo.gamercollection.databinding.FragmentRegisterBinding
-import es.upsa.mimo.gamercollection.extensions.afterTextChanged
-import es.upsa.mimo.gamercollection.extensions.clearErrors
-import es.upsa.mimo.gamercollection.extensions.onFocusChange
-import es.upsa.mimo.gamercollection.utils.Constants
+import es.upsa.mimo.gamercollection.extensions.doAfterTextChanged
+import es.upsa.mimo.gamercollection.extensions.getValue
+import es.upsa.mimo.gamercollection.extensions.setEndIconOnClickListener
+import es.upsa.mimo.gamercollection.extensions.setError
+import es.upsa.mimo.gamercollection.utils.StatusBarStyle
 import es.upsa.mimo.gamercollection.viewmodelfactories.RegisterViewModelFactory
 import es.upsa.mimo.gamercollection.viewmodels.RegisterViewModel
 
 class RegisterFragment : BindingFragment<FragmentRegisterBinding>() {
+
+    //region Protected properties
+    override val statusBarStyle = StatusBarStyle.SECONDARY
+    override val hasOptionsMenu = false
+    //endregion
 
     //region Private properties
     private lateinit var viewModel: RegisterViewModel
@@ -23,104 +29,90 @@ class RegisterFragment : BindingFragment<FragmentRegisterBinding>() {
     //region Lifecycle methods
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        initializeUI()
+        initializeUi()
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        binding.textInputLayoutUsername.doAfterTextChanged {
+            registerDataChanged()
+        }
+        binding.textInputLayoutPassword.doAfterTextChanged {
+            registerDataChanged()
+        }
+        binding.textInputLayoutConfirmPassword.doAfterTextChanged {
+            registerDataChanged()
+        }
     }
     //endregion
 
-    //region Private methods
-    private fun initializeUI() {
+    //region Public methods
+    fun register() {
+
+        binding.textInputLayoutUsername.textInputEditText.clearFocus()
+        binding.textInputLayoutPassword.textInputEditText.clearFocus()
+        binding.textInputLayoutConfirmPassword.textInputEditText.clearFocus()
+        viewModel.register(
+            binding.textInputLayoutUsername.getValue(),
+            binding.textInputLayoutPassword.getValue()
+        )
+    }
+    //endregion
+
+    //region Protected methods
+    override fun initializeUi() {
+        super.initializeUi()
 
         val application = activity?.application
         viewModel = ViewModelProvider(
             this,
             RegisterViewModelFactory(application)
-        ).get(RegisterViewModel::class.java)
+        )[RegisterViewModel::class.java]
         setupBindings()
 
-        with(binding) {
-
-            editTextUser.afterTextChanged {
-                registerDataChanged()
-            }
-            editTextUser.onFocusChange {
-                registerDataChanged()
-            }
-
-            imageButtonInfo.setOnClickListener {
-                showPopupDialog(resources.getString(R.string.username_info))
-            }
-
-            editTextPassword.afterTextChanged {
-                registerDataChanged()
-            }
-            editTextPassword.onFocusChange {
-                registerDataChanged()
-            }
-
-            imageButtonPassword.setOnClickListener {
-                Constants.showOrHidePassword(
-                    editTextPassword,
-                    imageButtonPassword,
-                    Constants.isDarkMode(context)
-                )
-            }
-
-            editTextRepeatPassword.afterTextChanged {
-                registerDataChanged()
-            }
-            editTextRepeatPassword.onFocusChange {
-                registerDataChanged()
-            }
-
-            imageButtonConfirmPassword.setOnClickListener {
-                Constants.showOrHidePassword(
-                    editTextRepeatPassword,
-                    imageButtonConfirmPassword,
-                    Constants.isDarkMode(context)
-                )
-            }
-
-            registerButton.setOnClickListener {
-                register()
-            }
+        binding.textInputLayoutUsername.setEndIconOnClickListener {
+            showPopupDialog(resources.getString(R.string.username_info))
         }
+        binding.fragment = this
+        binding.viewModel = viewModel
+        binding.lifecycleOwner = this
     }
+    //endregion
 
+    //region Private methods
     private fun setupBindings() {
 
-        viewModel.registerFormState.observe(viewLifecycleOwner, {
+        viewModel.registerFormState.observe(viewLifecycleOwner) {
 
             val registerState = it ?: return@observe
 
             with(binding) {
-
-                editTextUser.clearErrors()
-                editTextPassword.clearErrors()
-                editTextRepeatPassword.clearErrors()
-
-                registerButton.isEnabled = registerState.isDataValid
+                textInputLayoutUsername.setError("")
+                textInputLayoutPassword.setError("")
+                textInputLayoutConfirmPassword.setError("")
 
                 if (registerState.usernameError != null) {
-                    editTextUser.error = getString(registerState.usernameError)
+                    textInputLayoutUsername.setError(getString(registerState.usernameError))
                 }
                 if (registerState.passwordError != null) {
 
-                    editTextPassword.error = getString(registerState.passwordError)
-                    editTextRepeatPassword.error = getString(registerState.passwordError)
+                    textInputLayoutPassword.setError(getString(registerState.passwordError))
+                    textInputLayoutConfirmPassword.setError(getString(registerState.passwordError))
                 }
             }
-        })
+        }
 
-        viewModel.registerLoading.observe(viewLifecycleOwner, { isLoading ->
+        viewModel.registerLoading.observe(viewLifecycleOwner) { isLoading ->
 
             if (isLoading) {
                 showLoading()
             } else {
                 hideLoading()
             }
-        })
+        }
 
-        viewModel.registerError.observe(viewLifecycleOwner, { error ->
+        viewModel.registerError.observe(viewLifecycleOwner) { error ->
 
             if (error == null) {
                 launchActivity(MainActivity::class.java)
@@ -129,35 +121,16 @@ class RegisterFragment : BindingFragment<FragmentRegisterBinding>() {
                 hideLoading()
                 manageError(error)
             }
-        })
+        }
     }
 
     private fun registerDataChanged() {
 
         viewModel.registerDataChanged(
-            binding.editTextUser.text.toString(),
-            binding.editTextPassword.text.toString(),
-            binding.editTextRepeatPassword.text.toString()
+            binding.textInputLayoutUsername.getValue(),
+            binding.textInputLayoutPassword.getValue(),
+            binding.textInputLayoutConfirmPassword.getValue()
         )
-    }
-
-    private fun register() {
-
-        val username = binding.editTextUser.text.toString()
-        val password = binding.editTextPassword.text.toString()
-        val repeatPassword = binding.editTextRepeatPassword.text.toString()
-
-        if (username.isEmpty() || password.isEmpty() || repeatPassword.isEmpty()) {
-            showPopupDialog(resources.getString(R.string.error_registration_empty_data))
-            return
-        }
-
-        if (password != repeatPassword) {
-            showPopupDialog(resources.getString(R.string.error_registration_different_passwords))
-            return
-        }
-
-        viewModel.register(username, password)
     }
     //endregion
 }
