@@ -1,26 +1,11 @@
 package es.upsa.mimo.gamercollection.fragments.popups
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
+import es.upsa.mimo.gamercollection.base.BindingDialogFragment
+import es.upsa.mimo.gamercollection.databinding.DialogFragmentPopupLoadingBinding
 
-import es.upsa.mimo.gamercollection.R
+class PopupLoadingDialogFragment : BindingDialogFragment<DialogFragmentPopupLoadingBinding>() {
 
-class PopupLoadingDialogFragment : DialogFragment() {
-
-    //region Lifecycle methods
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setStyle(STYLE_NO_TITLE, R.style.Theme_GamerCollection_DialogTransparent)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_popup_loading_dialog, container, false)
-    }
+    //region Protected properties
+    override val transparentStyle = true
     //endregion
 }
