@@ -1,5 +1,7 @@
 package es.upsa.mimo.gamercollection.models
 
+import es.upsa.mimo.gamercollection.extensions.toString
+import es.upsa.mimo.gamercollection.utils.SharedPreferencesHelper
 import java.util.*
 
 data class FilterModel(
@@ -18,4 +20,37 @@ data class FilterModel(
     val isLoaned: Boolean,
     val hasSaga: Boolean,
     val hasSongs: Boolean
-)
+) {
+
+    fun minReleaseDateAsHumanReadable(): String? {
+
+        return minReleaseDate.toString(
+            SharedPreferencesHelper.filterDateFormat,
+            SharedPreferencesHelper.language
+        )
+    }
+
+    fun maxReleaseDateAsHumanReadable(): String? {
+
+        return maxReleaseDate.toString(
+            SharedPreferencesHelper.filterDateFormat,
+            SharedPreferencesHelper.language
+        )
+    }
+
+    fun minPurchaseDateAsHumanReadable(): String? {
+
+        return minPurchaseDate.toString(
+            SharedPreferencesHelper.filterDateFormat,
+            SharedPreferencesHelper.language
+        )
+    }
+
+    fun maxPurchaseDateAsHumanReadable(): String? {
+
+        return maxPurchaseDate.toString(
+            SharedPreferencesHelper.filterDateFormat,
+            SharedPreferencesHelper.language
+        )
+    }
+}
