@@ -1,12 +1,16 @@
 package es.upsa.mimo.gamercollection.repositories
 
+import es.upsa.mimo.gamercollection.R
 import es.upsa.mimo.gamercollection.injection.modules.MainDispatcher
+import es.upsa.mimo.gamercollection.models.login.UserData
 import es.upsa.mimo.gamercollection.models.requests.LoginCredentials
 import es.upsa.mimo.gamercollection.models.requests.NewPassword
 import es.upsa.mimo.gamercollection.models.responses.ErrorResponse
 import es.upsa.mimo.gamercollection.network.ApiManager
 import es.upsa.mimo.gamercollection.network.RequestResult
 import es.upsa.mimo.gamercollection.network.UserApiService
+import es.upsa.mimo.gamercollection.utils.Constants
+import es.upsa.mimo.gamercollection.utils.SharedPreferencesHelper
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
@@ -29,14 +33,35 @@ class UserRepository @Inject constructor(
         success: (String) -> Unit,
         failure: (ErrorResponse) -> Unit
     ) {
-        externalScope.launch {
-
-            val body = LoginCredentials(username, password)
-            when (val response = ApiManager.validateResponse(api.login(body))) {
-                is RequestResult.JsonSuccess -> success(response.body.token)
-                is RequestResult.Failure -> failure(response.error)
-            }
+        val userData = SharedPreferencesHelper.userData
+        if (username == Constants.GOOGLE_USER_TEST && password == Constants.GOOGLE_PASSWORD_TEST) {
+            SharedPreferencesHelper.userData = UserData(
+                Constants.GOOGLE_USER_TEST,
+                Constants.GOOGLE_PASSWORD_TEST,
+                false
+            )
+            success("-")
+        } else if (userData.username.isEmpty() || userData.username != username) {
+            failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.username_not_exist))
+        } else if (userData.username == username && userData.password == password) {
+            success("-")
+        } else {
+            failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.wrong_password))
         }
+//        externalScope.launch {
+//
+//            val body = LoginCredentials(username, password)
+//
+//            try {
+//                when (val response = ApiManager.validateResponse(api.login(body))) {
+//                    is RequestResult.JsonSuccess -> success(response.body.token)
+//                    is RequestResult.Failure -> failure(response.error)
+//                    else -> failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server))
+//                }
+//            } catch (e: Exception) {
+//                failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server_connection))
+//            }
+//        }
     }
 
     fun register(
@@ -45,41 +70,65 @@ class UserRepository @Inject constructor(
         success: () -> Unit,
         failure: (ErrorResponse) -> Unit
     ) {
-        externalScope.launch {
-
-            val body = LoginCredentials(username, password)
-            when (val response = ApiManager.validateResponse(api.register(body))) {
-                is RequestResult.Success -> success()
-                is RequestResult.Failure -> failure(response.error)
-            }
-        }
+        SharedPreferencesHelper.userData = UserData(username, password, false)
+        success()
+//        externalScope.launch {
+//
+//            val body = LoginCredentials(username, password)
+//
+//            try {
+//                when (val response = ApiManager.validateResponse(api.register(body))) {
+//                    is RequestResult.Success -> success()
+//                    is RequestResult.Failure -> failure(response.error)
+//                    else -> failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server))
+//                }
+//            } catch (e: Exception) {
+//                failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server_connection))
+//            }
+//        }
     }
 
     fun logout() {
         externalScope.launch {
-            api.logout()
+
+            try {
+                api.logout()
+            } catch (e: Exception) {
+            }
         }
     }
 
     fun updatePassword(password: String, success: () -> Unit, failure: (ErrorResponse) -> Unit) {
-        externalScope.launch {
-
-            val body = NewPassword(password)
-            when (val response = ApiManager.validateResponse(api.updatePassword(body))) {
-                is RequestResult.Success -> success()
-                is RequestResult.Failure -> failure(response.error)
-            }
-        }
+        success()
+//        externalScope.launch {
+//
+//            try {
+//                val body = NewPassword(password)
+//                when (val response = ApiManager.validateResponse(api.updatePassword(body))) {
+//                    is RequestResult.Success -> success()
+//                    is RequestResult.Failure -> failure(response.error)
+//                    else -> failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server))
+//                }
+//            } catch (e: Exception) {
+//                failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server_connection))
+//            }
+//        }
     }
 
     fun deleteUser(success: () -> Unit, failure: (ErrorResponse) -> Unit) {
-        externalScope.launch {
-
-            when (val response = ApiManager.validateResponse(api.deleteUser())) {
-                is RequestResult.Success -> success()
-                is RequestResult.Failure -> failure(response.error)
-            }
-        }
+        success()
+//        externalScope.launch {
+//
+//            try {
+//                when (val response = ApiManager.validateResponse(api.deleteUser())) {
+//                    is RequestResult.Success -> success()
+//                    is RequestResult.Failure -> failure(response.error)
+//                    else -> failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server))
+//                }
+//            } catch (e: Exception) {
+//                failure(ErrorResponse(Constants.EMPTY_VALUE, R.string.error_server_connection))
+//            }
+//        }
     }
     //endregion
 }
