@@ -1,23 +1,10 @@
 package es.upsa.mimo.gamercollection.utils
 
-import android.annotation.SuppressLint
-import android.app.Activity
-import android.content.Context
-import android.content.res.Configuration
-import android.graphics.drawable.Drawable
-import android.text.method.HideReturnsTransformationMethod
-import android.text.method.PasswordTransformationMethod
-import android.util.Log
-import android.view.inputmethod.InputMethodManager
-import android.widget.EditText
-import android.widget.ImageButton
-import androidx.core.content.ContextCompat
 import com.google.android.gms.maps.model.LatLng
-import es.upsa.mimo.gamercollection.R
-import es.upsa.mimo.gamercollection.adapters.SpinnerAdapter
-import java.text.DecimalFormat
-import java.text.SimpleDateFormat
-import java.util.*
+import es.upsa.mimo.gamercollection.models.FormatModel
+import es.upsa.mimo.gamercollection.models.GenreModel
+import es.upsa.mimo.gamercollection.models.PlatformModel
+import es.upsa.mimo.gamercollection.models.StateModel
 
 object Preferences {
     const val PREFERENCES_NAME = "preferences"
@@ -27,152 +14,87 @@ object Preferences {
     const val LANGUAGE_PREFERENCES_NAME = "language"
     const val ENGLISH_LANGUAGE_KEY = "en"
     const val SPANISH_LANGUAGE_KEY = "es"
-    const val SORTING_KEY_PREFERENCES_NAME = "sortingKey"
-    const val DEFAULT_SORTING_KEY = "name"
+    const val SORT_PARAM_PREFERENCES_NAME = "sortParam"
+    const val DEFAULT_SORT_PARAM = "name"
     const val SWIPE_REFRESH_PREFERENCES_NAME = "swipeRefreshEnabled"
     const val GAME_NOTIFICATION_PREFERENCES_NAME = "gameNotificationLaunched_"
-    const val VERSION_PREFERENCE_NAME = "version"
-    const val THEME_MODE_PREFERENCE_NAME = "themeMode"
+    const val THEME_MODE_PREFERENCES_NAME = "themeMode"
+    const val SORT_ORDER_PREFERENCE_NAME = "sortOrder"
+    const val NEW_CHANGES_POPUP_PREFERENCES_NAME = "newChanges"
 }
 
 object Constants {
-    const val POPUP_DIALOG = "popupDialog"
     const val LOADING_DIALOG = "loadingDialog"
     const val SYNC_DIALOG = "syncDialog"
-    const val EMPTY_VALUE = ""
     const val POINT_UP = 0f
     const val POINT_DOWN = -180f
-    const val NEXT_VALUE_SEPARATOR = ", "
     const val DATABASE_NAME = "GamerCollection"
-
-    fun getAdapter(
-        context: Context,
-        data: List<String>,
-        firstOptionEnabled: Boolean = false
-    ): SpinnerAdapter {
-
-        val arrayAdapter = SpinnerAdapter(context, data, firstOptionEnabled)
-        arrayAdapter.setDropDownViewResource(R.layout.spinner_dropdown_item)
-        return arrayAdapter
-    }
-
-    fun isDarkMode(context: Context?): Boolean {
-
-        val mode =
-            context?.resources?.configuration?.uiMode?.and(Configuration.UI_MODE_NIGHT_MASK)
-        return mode == Configuration.UI_MODE_NIGHT_YES
-    }
-
-    fun hideSoftKeyboard(activity: Activity) {
-
-        activity.currentFocus?.let { currentFocus ->
-
-            val inputMethodManager =
-                activity.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
-            inputMethodManager.hideSoftInputFromWindow(currentFocus.windowToken, 0)
-        } ?: return
-    }
-
     const val DATE_FORMAT = "yyyy-MM-dd"
-
-    fun getDateFormatToShow(language: String): String {
-
-        return when (language) {
-            Preferences.SPANISH_LANGUAGE_KEY -> "d MMMM yyyy"
-            else -> "MMMM d, yyyy"
-        }
-    }
-
-    fun getFilterDateFormat(language: String): String {
-
-        return when (language) {
-            Preferences.SPANISH_LANGUAGE_KEY -> "dd/MM/yyyy"
-            else -> "MM/dd/yyyy"
-        }
-    }
-
-    @SuppressLint("SimpleDateFormat")
-    fun dateToString(
-        date: Date?,
-        format: String? = null,
-        language: String? = null
-    ): String? {
-
-        val dateFormat = format ?: DATE_FORMAT
-        val locale = language?.let {
-            Locale.forLanguageTag(it)
-        } ?: run {
-            Locale.getDefault()
-        }
-        date?.let {
-
-            return try {
-                SimpleDateFormat(dateFormat, locale).format(it)
-            } catch (e: Exception) {
-
-                Log.e("Constants", e.message ?: EMPTY_VALUE)
-                null
-            }
-        } ?: run {
-
-            Log.e("Constants", "date null")
-            return null
-        }
-    }
-
-    @SuppressLint("SimpleDateFormat")
-    fun stringToDate(
-        dateString: String?,
-        format: String? = null,
-        language: String? = null
-    ): Date? {
-
-        val dateFormat = format ?: DATE_FORMAT
-        val locale = language?.let {
-            Locale.forLanguageTag(it)
-        } ?: run {
-            Locale.getDefault()
-        }
-        dateString?.let {
-
-            return try {
-                SimpleDateFormat(dateFormat, locale).parse(it)
-            } catch (e: Exception) {
-
-                Log.e("Constants", e.message ?: EMPTY_VALUE)
-                null
-            }
-        } ?: run {
-            Log.e("Constants", "dateString null")
-            return null
-        }
-    }
-
-    fun getFormattedNumber(number: Int): String {
-
-        val formatter = DecimalFormat("#,###")
-        return formatter.format(number)
-    }
-
-    const val GAME_ID = "gameId"
-    const val IS_RAWG_GAME = "isRawgGame"
+    const val EMPTY_VALUE = ""
     const val NO_VALUE = "-"
-
-    fun getPegiImage(pegi: String?, context: Context): Drawable? {
-
-        return when (pegi) {
-            "+3" -> ContextCompat.getDrawable(context, R.drawable.pegi3)
-            "+7" -> ContextCompat.getDrawable(context, R.drawable.pegi7)
-            "+12" -> ContextCompat.getDrawable(context, R.drawable.pegi12)
-            "+16" -> ContextCompat.getDrawable(context, R.drawable.pegi16)
-            "+18" -> ContextCompat.getDrawable(context, R.drawable.pegi18)
-            else -> null
-        }
-    }
-
-    const val SAGA_ID = "sagaId"
-
+    const val NEXT_VALUE_SEPARATOR = ", "
     val DEFAULT_LOCATION = LatLng(40.4169019, -3.7056721)
+    const val GOOGLE_USER_TEST = "googleTest"
+    const val GOOGLE_PASSWORD_TEST = "d9MqzK3k1&07"
+
+    var FORMATS = listOf(
+        FormatModel("DIGITAL", "Digital"),
+        FormatModel("PHYSICAL", "Physical"),
+        FormatModel("OTHER", "Other")
+    )
+    var GENRES = listOf(
+        GenreModel("ACTION", "Action"),
+        GenreModel("ACTION_ADVENTURE", "Action-Adventure"),
+        GenreModel("ADVENTURE", "Adventure"),
+        GenreModel("ARCADE", "Arcade"),
+        GenreModel("FIGHTING", "Fighting"),
+        GenreModel("HACK_AND_SLASH", "Hack and Slash"),
+        GenreModel("OTHER", "Other"),
+        GenreModel("PLATFORMS", "Platforms"),
+        GenreModel("PUZZLE", "Puzzle"),
+        GenreModel("RACING", "Racing"),
+        GenreModel("ROL", "Rol"),
+        GenreModel("SHOOTER", "Shooter"),
+        GenreModel("SIMULATOR", "Simulator"),
+        GenreModel("SOCIAL", "Social"),
+        GenreModel("SPORTS", "Sports"),
+        GenreModel("STRATEGY", "Strategy"),
+        GenreModel("SURVIVIAL_HORROR", "Survival Horror")
+    )
+    var PLATFORMS = listOf(
+        PlatformModel("PSONE", "PlayStation"),
+        PlatformModel("PS2", "PlayStation 2"),
+        PlatformModel("PS3", "PlayStation 3"),
+        PlatformModel("PS4", "PlayStation 4"),
+        PlatformModel("PS5", "PlayStation 5"),
+        PlatformModel("PSP", "PlayStation Portable"),
+        PlatformModel("PSVITA", "PlayStation Vita"),
+        PlatformModel("XBOX", "Xbox"),
+        PlatformModel("XBOX_360", "Xbox 360"),
+        PlatformModel("XBOX_ONE", "Xbox One"),
+        PlatformModel("XBOX_SERIES_X", "Xbox Series X"),
+        PlatformModel("PC", "PC"),
+        PlatformModel("SEGA", "Sega"),
+        PlatformModel("NES", "NES"),
+        PlatformModel("SUPER_NINTENDO", "Super Nintendo"),
+        PlatformModel("NINTENDO_64", "Nintendo 64"),
+        PlatformModel("GAME_CUBE", "Game Cube"),
+        PlatformModel("GAME_BOY", "Game Boy"),
+        PlatformModel("GBC", "Game Boy Color"),
+        PlatformModel("GBA", "Game Boy Advance"),
+        PlatformModel("GBA_SP", "Game Boy Advance SP"),
+        PlatformModel("NINTENDO_DS", "Nintendo DS"),
+        PlatformModel("NINTENDO_3DS", "Nintendo 3DS"),
+        PlatformModel("WII", "Wii"),
+        PlatformModel("WII_U", "Wii U"),
+        PlatformModel("NINTENDO_SWITCH", "Nintendo Switch"),
+        PlatformModel("OTHER", "Other")
+    )
+    var STATES = listOf(
+        StateModel("PENDING", "Pending"),
+        StateModel("IN_PROGRESS", "In Progress"),
+        StateModel("FINISHED", "Finished")
+    )
 
     fun isUserNameValid(username: String): Boolean {
         return username.isNotBlank()
@@ -180,23 +102,6 @@ object Constants {
 
     fun isPasswordValid(password: String): Boolean {
         return password.length > 3
-    }
-
-    fun showOrHidePassword(editText: EditText, imageButton: ImageButton, isDarkMode: Boolean) {
-
-        if (editText.transformationMethod is HideReturnsTransformationMethod) {
-
-            val image =
-                if (isDarkMode) R.drawable.ic_show_password_dark else R.drawable.ic_show_password_light
-            imageButton.setImageResource(image)
-            editText.transformationMethod = PasswordTransformationMethod.getInstance()
-        } else {
-
-            val image =
-                if (isDarkMode) R.drawable.ic_hide_password_dark else R.drawable.ic_hide_password_light
-            imageButton.setImageResource(image)
-            editText.transformationMethod = HideReturnsTransformationMethod.getInstance()
-        }
     }
 }
 
@@ -209,4 +114,34 @@ object State {
 object Notifications {
     const val CHANNEL_ID = "GAMER_COLLECTION_NOTIFICATIONS_CHANNEL_ID"
     const val CHANNEL_GROUP = "GAMER_COLLECTION_NOTIFICATIONS_CHANNEL_GROUP"
+}
+
+enum class ScrollPosition {
+    TOP, MIDDLE, END
+}
+
+enum class StatusBarStyle {
+    PRIMARY,
+    SECONDARY
+}
+
+enum class CustomInputType {
+    TEXT,
+    MULTI_LINE_TEXT,
+    NUMBER,
+    PASSWORD,
+    URL,
+    DATE,
+    NONE
+}
+
+enum class CustomDropdownType {
+    FORMAT,
+    GENRE,
+    PEGI,
+    PLATFORM,
+    STATE,
+    SORT_PARAM,
+    SORT_ORDER,
+    APP_THEME
 }
