@@ -2,9 +2,9 @@ package es.upsa.mimo.gamercollection.viewholders
 
 import androidx.recyclerview.widget.RecyclerView
 import es.upsa.mimo.gamercollection.adapters.OnItemClickListener
-import es.upsa.mimo.gamercollection.databinding.LayoutLoadMoreItemsBinding
+import es.upsa.mimo.gamercollection.databinding.ItemLoadMoreItemsBinding
 
-class LoadMoreItemsViewHolder(val binding: LayoutLoadMoreItemsBinding) :
+class LoadMoreItemsViewHolder(val binding: ItemLoadMoreItemsBinding) :
     RecyclerView.ViewHolder(binding.root) {
 
     //region Public methods
