@@ -5,22 +5,18 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import es.upsa.mimo.gamercollection.GamerCollectionApplication
 import es.upsa.mimo.gamercollection.repositories.GameRepository
-import es.upsa.mimo.gamercollection.repositories.PlatformRepository
 import es.upsa.mimo.gamercollection.repositories.SagaRepository
 import es.upsa.mimo.gamercollection.viewmodels.SagaDetailViewModel
 import javax.inject.Inject
 
 class SagaDetailViewModelFactory(
     private val application: Application?,
-    private val sagaId: Int?
+    private val sagaId: Int
 ) : ViewModelProvider.Factory {
 
     //region Public properties
     @Inject
     lateinit var gameRepository: GameRepository
-
-    @Inject
-    lateinit var platformRepository: PlatformRepository
 
     @Inject
     lateinit var sagaRepository: SagaRepository
