@@ -3,18 +3,15 @@ package es.upsa.mimo.gamercollection.viewmodels
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import es.upsa.mimo.gamercollection.fragments.GamesFragment.ScrollPosition
+import es.upsa.mimo.gamercollection.utils.ScrollPosition
 import es.upsa.mimo.gamercollection.models.responses.ErrorResponse
 import es.upsa.mimo.gamercollection.models.responses.GameResponse
-import es.upsa.mimo.gamercollection.models.responses.PlatformResponse
 import es.upsa.mimo.gamercollection.repositories.GameRepository
-import es.upsa.mimo.gamercollection.repositories.PlatformRepository
 import es.upsa.mimo.gamercollection.utils.SharedPreferencesHelper
 import javax.inject.Inject
 
 class GameSearchViewModel @Inject constructor(
-    private val gameRepository: GameRepository,
-    private val platformRepository: PlatformRepository
+    private val gameRepository: GameRepository
 ) : ViewModel() {
 
     //region Private properties
@@ -22,16 +19,14 @@ class GameSearchViewModel @Inject constructor(
     private val _gamesLoading = MutableLiveData<Boolean>()
     private val _gamesError = MutableLiveData<ErrorResponse>()
     private val _games = MutableLiveData<MutableList<GameResponse>>()
-    private val _gamesCount = MutableLiveData<Int>()
+    private val _gamesCount = MutableLiveData(0)
     private val _scrollPosition = MutableLiveData(ScrollPosition.TOP)
     //endregion
 
     //region Public properties
     var query: String? = null
     val swipeRefresh: Boolean
-        get() = SharedPreferencesHelper.getSwipeRefresh()
-    val platforms: List<PlatformResponse>
-        get() = platformRepository.getPlatformsDatabase()
+        get() = SharedPreferencesHelper.swipeRefresh
     val gamesLoading: LiveData<Boolean> = _gamesLoading
     val gamesError: LiveData<ErrorResponse> = _gamesError
     val games: LiveData<MutableList<GameResponse>> = _games
@@ -69,14 +64,14 @@ class GameSearchViewModel @Inject constructor(
     private fun addGames(newGames: List<GameResponse>, next: Boolean) {
 
         val currentGames = _games.value ?: mutableListOf()
-        if (currentGames.size > 0) {
+        if (currentGames.isNotEmpty()) {
             currentGames.removeLast()
         }
         currentGames.addAll(newGames)
         if (next) {
             currentGames.add(
                 GameResponse(
-                    0,
+                    -1,
                     null,
                     null,
                     0.0,
@@ -97,7 +92,7 @@ class GameSearchViewModel @Inject constructor(
                     null,
                     null,
                     null,
-                    listOf()
+                    mutableListOf()
                 )
             )
         }
