@@ -1,6 +1,5 @@
 package es.upsa.mimo.gamercollection.viewmodels
 
-import android.app.AlertDialog
 import android.content.Context
 import android.content.res.Resources
 import android.view.Gravity
@@ -9,20 +8,20 @@ import android.widget.NumberPicker
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import es.upsa.mimo.gamercollection.R
-import es.upsa.mimo.gamercollection.fragments.GamesFragment
+import es.upsa.mimo.gamercollection.extensions.getPickerParams
+import es.upsa.mimo.gamercollection.extensions.setup
 import es.upsa.mimo.gamercollection.models.FilterModel
 import es.upsa.mimo.gamercollection.models.responses.ErrorResponse
 import es.upsa.mimo.gamercollection.models.responses.GameResponse
-import es.upsa.mimo.gamercollection.models.responses.PlatformResponse
 import es.upsa.mimo.gamercollection.repositories.GameRepository
-import es.upsa.mimo.gamercollection.repositories.PlatformRepository
+import es.upsa.mimo.gamercollection.utils.ScrollPosition
 import es.upsa.mimo.gamercollection.utils.SharedPreferencesHelper
 import javax.inject.Inject
 
 class GamesViewModel @Inject constructor(
-    private val gameRepository: GameRepository,
-    private val platformRepository: PlatformRepository
+    private val gameRepository: GameRepository
 ) : ViewModel() {
 
     //region Private properties
@@ -34,19 +33,21 @@ class GamesViewModel @Inject constructor(
     private val _gameDeleted = MutableLiveData<Int?>()
     private var _state = MutableLiveData<String?>(null)
     private var _filters = MutableLiveData<FilterModel?>(null)
-    private var _scrollPosition = MutableLiveData(GamesFragment.ScrollPosition.TOP)
-    private var sortKey: String = SharedPreferencesHelper.getSortingKey()
-    private var sortAscending = true
+    private var _scrollPosition = MutableLiveData(ScrollPosition.TOP)
+    private var sortParam: String = SharedPreferencesHelper.sortParam
+    private var isSortOrderAscending = SharedPreferencesHelper.isSortOrderAscending
     private var query: String? = null
     //endregion
 
     //region Public properties
     val language: String
-        get() = SharedPreferencesHelper.getLanguage()
+        get() = SharedPreferencesHelper.language
+    val dateFormatToShow: String
+        get() = SharedPreferencesHelper.dateFormatToShow
+    val filterDateFormat: String
+        get() = SharedPreferencesHelper.filterDateFormat
     val swipeRefresh: Boolean
-        get() = SharedPreferencesHelper.getSwipeRefresh()
-    val platforms: List<PlatformResponse>
-        get() = platformRepository.getPlatformsDatabase()
+        get() = SharedPreferencesHelper.swipeRefresh
     val gamesLoading: LiveData<Boolean> = _gamesLoading
     val gamesError: LiveData<ErrorResponse> = _gamesError
     val games: LiveData<List<GameResponse>> = _games
@@ -54,7 +55,7 @@ class GamesViewModel @Inject constructor(
     val gameDeleted: LiveData<Int?> = _gameDeleted
     val state: LiveData<String?> = _state
     val filters: LiveData<FilterModel?> = _filters
-    val scrollPosition: LiveData<GamesFragment.ScrollPosition> = _scrollPosition
+    val scrollPosition: LiveData<ScrollPosition> = _scrollPosition
     //endregion
 
     //region Public methods
@@ -68,7 +69,7 @@ class GamesViewModel @Inject constructor(
             _gamesLoading.value = false
         }, {
 
-            _gamesLoading.value = true
+            _gamesLoading.value = false
             _gamesError.value = it
         })
     }
@@ -78,8 +79,8 @@ class GamesViewModel @Inject constructor(
         val games = gameRepository.getGamesDatabase(
             _filters.value,
             query,
-            sortKey,
-            sortAscending
+            sortParam,
+            isSortOrderAscending
         )
         _originalGames.value = games
 
@@ -93,53 +94,40 @@ class GamesViewModel @Inject constructor(
 
         _gamesCount.value = games
 
-        _scrollPosition.value = GamesFragment.ScrollPosition.TOP
+        _scrollPosition.value = ScrollPosition.TOP
     }
 
     fun sortGames(context: Context, resources: Resources) {
 
-        val sortingKeys = resources.getStringArray(R.array.sorting_keys_ids)
-        val sortingValues = resources.getStringArray(R.array.sorting_keys)
+        val sortingKeys = resources.getStringArray(R.array.sort_param_keys)
+        val sortingValues = resources.getStringArray(R.array.sort_param_values)
 
         val dialogView = LinearLayout(context)
         dialogView.orientation = LinearLayout.HORIZONTAL
 
-        val ascendingPicker = getPicker(
-            arrayOf(
-                resources.getString(R.string.ascending),
-                resources.getString(R.string.descending)
-            ), context
-        )
-        ascendingPicker.value = if (sortAscending) 0 else 1
-        val ascendingPickerParams = LinearLayout.LayoutParams(
-            LinearLayout.LayoutParams.MATCH_PARENT,
-            LinearLayout.LayoutParams.WRAP_CONTENT
-        )
-        ascendingPickerParams.weight = 1f
+        val sortKeysPicker = NumberPicker(context)
+        sortKeysPicker.setup(sortingValues)
+        sortKeysPicker.value = sortingKeys.indexOf(sortParam)
 
-        val sortKeysPicker = getPicker(sortingValues, context)
-        sortKeysPicker.value = sortingKeys.indexOf(sortKey)
-        val sortKeysPickerParams = LinearLayout.LayoutParams(
-            LinearLayout.LayoutParams.MATCH_PARENT,
-            LinearLayout.LayoutParams.WRAP_CONTENT
-        )
-        sortKeysPickerParams.weight = 1f
+        val sortOrdersPicker = NumberPicker(context)
+        sortOrdersPicker.setup(context.resources.getStringArray(R.array.sort_order_values))
+        sortOrdersPicker.value = if (isSortOrderAscending) 0 else 1
 
         val params = LinearLayout.LayoutParams(50, 50)
         params.gravity = Gravity.CENTER
 
         dialogView.layoutParams = params
-        dialogView.addView(sortKeysPicker, sortKeysPickerParams)
-        dialogView.addView(ascendingPicker, ascendingPickerParams)
+        dialogView.addView(sortKeysPicker, getPickerParams())
+        dialogView.addView(sortOrdersPicker, getPickerParams())
 
-        AlertDialog.Builder(context)
+        MaterialAlertDialogBuilder(context)
             .setTitle(resources.getString(R.string.sort_title))
             .setView(dialogView)
             .setCancelable(false)
             .setPositiveButton(resources.getString(R.string.accept)) { dialog, _ ->
 
-                sortKey = sortingKeys[sortKeysPicker.value]
-                sortAscending = ascendingPicker.value == 0
+                sortParam = sortingKeys[sortKeysPicker.value]
+                isSortOrderAscending = sortOrdersPicker.value == 0
                 fetchGames()
                 dialog.dismiss()
             }
@@ -201,29 +189,18 @@ class GamesViewModel @Inject constructor(
         fetchGames()
     }
 
-    fun setPosition(newPosition: GamesFragment.ScrollPosition) {
+    fun setPosition(newPosition: ScrollPosition) {
         _scrollPosition.value = newPosition
     }
     //endregion
 
     //region Private methods
-    private fun getPicker(values: Array<String>, context: Context): NumberPicker {
-
-        val picker = NumberPicker(context)
-        picker.minValue = 0
-        picker.maxValue = values.size - 1
-        picker.wrapSelectorWheel = true
-        picker.descendantFocusability = NumberPicker.FOCUS_BLOCK_DESCENDANTS
-        picker.displayedValues = values
-        return picker
-    }
-
     private fun resetProperties() {
 
         _state.value = null
         _filters.value = null
-        sortKey = SharedPreferencesHelper.getSortingKey()
-        sortAscending = true
+        sortParam = SharedPreferencesHelper.sortParam
+        isSortOrderAscending = true
         query = null
     }
     //endregion
