@@ -11,12 +11,8 @@ import es.upsa.mimo.gamercollection.utils.SharedPreferencesHelper
 import javax.inject.Inject
 
 class LandingViewModel @Inject constructor(
-    private val formatRepository: FormatRepository,
     private val gameRepository: GameRepository,
-    private val genreRepository: GenreRepository,
-    private val platformRepository: PlatformRepository,
-    private val sagaRepository: SagaRepository,
-    private val stateRepository: StateRepository
+    private val sagaRepository: SagaRepository
 ) : ViewModel() {
 
     //region Private properties
@@ -25,38 +21,29 @@ class LandingViewModel @Inject constructor(
 
     //region Public properties
     val language: String
-        get() = SharedPreferencesHelper.getLanguage()
+        get() = SharedPreferencesHelper.language
+    val newChangesPopupShown: Boolean
+        get() = SharedPreferencesHelper.newChangesPopupShown
     val landingClassToStart = _landingClassToStart
     //endregion
 
     //region Public methods
-    fun checkVersion() {
-
-        val currentVersion = SharedPreferencesHelper.getVersion()
-        val newVersion = BuildConfig.VERSION_CODE
-        if (newVersion > currentVersion) {
-
-            SharedPreferencesHelper.setVersion(newVersion)
-            SharedPreferencesHelper.removePassword()
-            SharedPreferencesHelper.removeCredentials()
-            resetDatabase()
-            _landingClassToStart.value = LoginActivity::class.java
-        } else {
-
-            _landingClassToStart.value = if (SharedPreferencesHelper.isLoggedIn()) {
-                MainActivity::class.java
-            } else {
-                LoginActivity::class.java
-            }
-        }
-    }
-
     fun checkTheme() {
 
-        when (SharedPreferencesHelper.getThemeMode()) {
+        when (SharedPreferencesHelper.themeMode) {
             1 -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
             2 -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
             else -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM)
+        }
+    }
+
+    fun checkIsLoggedIn() {
+
+        SharedPreferencesHelper.newChangesPopupShown = true
+        _landingClassToStart.value = if (SharedPreferencesHelper.isLoggedIn) {
+            MainActivity::class.java
+        } else {
+            LoginActivity::class.java
         }
     }
     //endregion
@@ -64,12 +51,8 @@ class LandingViewModel @Inject constructor(
     //region Private methods
     private fun resetDatabase() {
 
-        formatRepository.resetTable()
         gameRepository.resetTable()
-        genreRepository.resetTable()
-        platformRepository.resetTable()
         sagaRepository.resetTable()
-        stateRepository.resetTable()
     }
     //endregion
 }
