@@ -8,18 +8,16 @@ import es.upsa.mimo.gamercollection.models.login.AuthData
 import es.upsa.mimo.gamercollection.models.login.LoginFormState
 import es.upsa.mimo.gamercollection.models.login.UserData
 import es.upsa.mimo.gamercollection.models.responses.ErrorResponse
-import es.upsa.mimo.gamercollection.repositories.*
+import es.upsa.mimo.gamercollection.repositories.GameRepository
+import es.upsa.mimo.gamercollection.repositories.SagaRepository
+import es.upsa.mimo.gamercollection.repositories.UserRepository
 import es.upsa.mimo.gamercollection.utils.Constants
 import es.upsa.mimo.gamercollection.utils.SharedPreferencesHelper
 import javax.inject.Inject
 
 class LoginViewModel @Inject constructor(
-    private val formatRepository: FormatRepository,
     private val gameRepository: GameRepository,
-    private val genreRepository: GenreRepository,
-    private val platformRepository: PlatformRepository,
     private val sagaRepository: SagaRepository,
-    private val stateRepository: StateRepository,
     private val userRepository: UserRepository
 ) : ViewModel() {
 
@@ -31,7 +29,7 @@ class LoginViewModel @Inject constructor(
 
     //region Public properties
     val username: String
-        get() = SharedPreferencesHelper.getUserData().username
+        get() = SharedPreferencesHelper.userData.username
     val loginFormState: LiveData<LoginFormState> = _loginForm
     val loginLoading: LiveData<Boolean> = _loginLoading
     val loginError: LiveData<ErrorResponse?> = _loginError
@@ -44,10 +42,9 @@ class LoginViewModel @Inject constructor(
         userRepository.login(username, password, { token ->
 
             val userData = UserData(username, password, false)
-            val authData = AuthData(token)
             SharedPreferencesHelper.run {
-                storeUserData(userData)
-                storeCredentials(authData)
+                this.userData = userData
+                this.credentials = AuthData(token)
             }
             loadContent(userData)
         }, {
@@ -76,36 +73,20 @@ class LoginViewModel @Inject constructor(
     //region Private methods
     private fun loadContent(userData: UserData) {
 
-        formatRepository.loadFormats({
-            gameRepository.loadGames({
-                genreRepository.loadGenres({
-                    platformRepository.loadPlatforms({
-                        sagaRepository.loadSagas({
-                            stateRepository.loadStates({
+//        gameRepository.loadGames({
+//            sagaRepository.loadSagas({
 
-                                userData.isLoggedIn = true
-                                SharedPreferencesHelper.storeUserData(userData)
+                userData.isLoggedIn = true
+                SharedPreferencesHelper.userData = userData
 
-                                _loginError.value = null
-                                _loginLoading.value = false
-                            }, {
-                                _loginError.value = it
-                            })
-                        }, {
-                            _loginError.value = it
-                        })
-                    }, {
-                        _loginError.value = it
-                    })
-                }, {
-                    _loginError.value = it
-                })
-            }, {
-                _loginError.value = it
-            })
-        }, {
-            _loginError.value = it
-        })
+                _loginError.value = null
+                _loginLoading.value = false
+//            }, {
+//                _loginError.value = it
+//            })
+//        }, {
+//            _loginError.value = it
+//        })
     }
     //endregion
 }
